Reorganize error definitions
