+++ server/errors/BadRequestError.ts
@@ -1,0 +1,8 @@
+import { InternalError } from "./InternalError.js";
+
+export class BadRequestError extends InternalError {
+	constructor(message: string = "Invalid data") {
+		super({ status: 400, message, harmless: true });
+		this.name = "BadRequestError";
+	}
+}
