+++ server/errors/index.ts
@@ -1,0 +1,9 @@
+export * from "./BadMethodError.js";
+export * from "./BadRequestError.js";
+export * from "./DuplicateAccountError.js";
+export * from "./InternalError.js";
+export * from "./NotEnoughRoomError.js";
+export * from "./NotFoundError.js";
+export * from "./OriginError.js";
+export * from "./ThrottledError.js";
+export * from "./UnauthorizedError.js";
