+++ server/handleErrors.ts
@@ -1,5 +1,6 @@
 import type { ErrorRequestHandler } from "express";
-import { InternalError, respondError, respondInternalError } from "./responses.js";
+import { InternalError } from "./errors/index.js";
+import { respondError, respondInternalError } from "./responses.js";
 
 export const handleErrors: ErrorRequestHandler = (err: unknown, req, res, next) => {
 	if (res.headersSent) {
